dapp-update: reorganized sharedMethods and sharedComponents
