+++ system-contract-dapp-playground/__tests__/hedera/helper/index.test.ts
@@ -19,7 +19,10 @@
  */
 
 import { constructIHederaTokenKey } from '@/utils/contract-interactions/HTS/helpers';
-import { DEFAULT_IHTS_KEY_VALUE, KEY_TYPE_MAP } from '@/utils/contract-interactions/HTS/constant';
+import {
+  DEFAULT_IHTS_KEY_VALUE,
+  KEY_TYPE_MAP,
+} from '@/utils/contract-interactions/HTS/token-create-custom/constant';
 import {
   IHederaTokenServiceKeyType,
   IHederaTokenServiceKeyValueType,
