+++ system-contract-dapp-playground/src/components/contract-interaction/hts/shared/components/TransactionResultTable.tsx
@@ -18,379 +18,26 @@
  *
  */
 
+import Link from 'next/link';
 import { ethers } from 'ethers';
-import Image from 'next/image';
-import { AiOutlineMinus } from 'react-icons/ai';
 import { FiExternalLink } from 'react-icons/fi';
+import { AiOutlineMinus } from 'react-icons/ai';
 import { Dispatch, SetStateAction } from 'react';
 import { MdNavigateBefore, MdNavigateNext } from 'react-icons/md';
-import { convertCalmelCaseFunctionName } from '@/utils/common/helpers';
+import { TransactionResult } from '@/types/contract-interactions/HTS';
 import {
-  Tooltip,
-  Input,
-  Select,
-  Link,
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-  Table,
   TableContainer,
-  Tbody,
-  Td,
-  Th,
+  Table,
   Thead,
   Tr,
+  Th,
+  Tbody,
+  Td,
+  Popover,
+  PopoverTrigger,
+  Tooltip,
+  PopoverContent,
 } from '@chakra-ui/react';
-import {
-  handleAddingOrRemovingKeys,
-  handleKeyTypeOnChange,
-  handleKeyValueTypeOnChange,
-  handleUpdateKeyValue,
-} from './sharedMethods';
-import {
-  CommonKeyObject,
-  IHederaTokenServiceKeyType,
-  IHederaTokenServiceKeyValueType,
-  TransactionResult,
-} from '@/types/contract-interactions/HTS';
-
-/** @dev shared form input component */
-interface SharedFormInputFieldPageProps {
-  param: string;
-  paramKey: string;
-  paramType: string;
-  paramSize: string;
-  paramValue: string;
-  isDisable?: boolean;
-  explanation: string;
-  paramClassName: string;
-  paramFocusColor: string;
-  paramPlaceholder: string;
-  handleInputOnChange: (e: any, param: string) => void;
-}
-
-export const SharedFormInputField = ({
-  param,
-  paramKey,
-  paramType,
-  isDisable,
-  paramSize,
-  paramValue,
-  explanation,
-  paramClassName,
-  paramFocusColor,
-  paramPlaceholder,
-  handleInputOnChange,
-}: SharedFormInputFieldPageProps) => {
-  return (
-    <Tooltip key={paramKey} label={explanation} placement="top" fontWeight={'medium'}>
-      <Input
-        value={paramValue}
-        disabled={isDisable}
-        type={paramType}
-        onChange={(e) => handleInputOnChange(e, param)}
-        placeholder={paramPlaceholder}
-        size={paramSize}
-        focusBorderColor={paramFocusColor}
-        className={paramClassName}
-      />
-    </Tooltip>
-  );
-};
-
-/** @dev shared form button component */
-interface SharedFormButtonPageProps {
-  switcher: boolean;
-  explanation: string;
-  buttonTitle: string;
-  handleButtonOnClick: any;
-}
-
-export const SharedFormButton = ({
-  switcher,
-  explanation,
-  buttonTitle,
-  handleButtonOnClick,
-}: SharedFormButtonPageProps) => {
-  return (
-    <Tooltip label={explanation} placement="top" fontWeight={'medium'}>
-      <button
-        onClick={handleButtonOnClick}
-        className={`border w-full rounded-md py-1 text-lg transition duration-300 ${
-          switcher
-            ? ` border-hedera-purple text-hedera-purple`
-            : `border-white/30 hover:text-white hover:border-white`
-        }`}
-      >
-        <p>{buttonTitle}</p>
-      </button>
-    </Tooltip>
-  );
-};
-
-/** @dev shared execute button component */
-interface SharedExecuteButtonPageProps {
-  isLoading: boolean;
-  buttonTitle: string;
-  handleCreatingFungibleToken: () => Promise<void>;
-}
-
-export const SharedExecuteButton = ({
-  isLoading,
-  buttonTitle,
-  handleCreatingFungibleToken,
-}: SharedExecuteButtonPageProps) => {
-  return (
-    <button
-      onClick={handleCreatingFungibleToken}
-      disabled={isLoading}
-      className={`w-full border py-2 rounded-xl transition duration-300 ${
-        isLoading
-          ? 'cursor-not-allowed border-white/30 text-white/30'
-          : 'border-button-stroke-violet text-button-stroke-violet hover:bg-button-stroke-violet/60 hover:text-white'
-      }`}
-    >
-      {isLoading ? (
-        <div className="flex gap-1 justify-center">
-          Executing...
-          <Image
-            src={'/brandings/hedera-logomark.svg'}
-            alt={'hedera-logomark'}
-            width={15}
-            height={15}
-            className="animate-bounce"
-          />
-        </div>
-      ) : (
-        <>{buttonTitle}</>
-      )}
-    </button>
-  );
-};
-
-/** @dev shared execute button with service fee component */
-interface SharedExecuteButtonWithServiceFeePageProps {
-  isLoading: boolean;
-  paramValues: string;
-  executeBtnTitle: string;
-  handleCreatingFungibleToken: () => Promise<void>;
-  handleInputOnChange: (e: any, param: string) => void;
-}
-
-export const SharedExecuteButtonWithServiceFee = ({
-  isLoading,
-  paramValues,
-  executeBtnTitle,
-  handleInputOnChange,
-  handleCreatingFungibleToken,
-}: SharedExecuteButtonWithServiceFeePageProps) => {
-  return (
-    <div className="flex flex-col gap-1">
-      <div className="flex items-center gap-6">
-        {/* Service Fee */}
-        <div className="w-3/12">
-          <SharedFormInputField
-            param={'msgValue'}
-            paramValue={paramValues}
-            handleInputOnChange={handleInputOnChange}
-            paramSize={'lg'}
-            paramType={'number'}
-            paramKey={'msgValue'}
-            explanation={
-              'Represents the fee in HBAR directly paid to the contract system of the Hedera Token Service'
-            }
-            paramClassName={'border-white/30 rounded-xl'}
-            paramPlaceholder={'Service fee...'}
-            paramFocusColor={'#A98DF4'}
-          />
-        </div>
-        {/* Execute button */}
-        <div className=" w-9/12">
-          <SharedExecuteButton
-            isLoading={isLoading}
-            buttonTitle={executeBtnTitle}
-            handleCreatingFungibleToken={handleCreatingFungibleToken}
-          />
-        </div>
-      </div>
-      <p className="text-sm whitespace-normal">
-        <span className="italic font-medium text-sm">*Important:</span> Varying configurations
-        applied to the token will result in varying service fees. Be sure to utilize the{' '}
-        <Link
-          className="underline text-hedera-green font-medium whitespace-nowrap"
-          href={'https://hedera.com/fees'}
-          target="_blank"
-        >
-          Hedera service fee calculator
-        </Link>{' '}
-        for precise estimation of the applicable fee, as this fee is non-refundable.
-      </p>
-    </div>
-  );
-};
-
-/** @dev shared component presenting signing keys*/
-interface SigningKeyPageProps {
-  keys: CommonKeyObject[];
-  chosenKeys: Set<IHederaTokenServiceKeyType>;
-  keyTypesToShow: Set<IHederaTokenServiceKeyType>;
-  HederaTokenKeyTypes: IHederaTokenServiceKeyType[];
-  setKeys: Dispatch<SetStateAction<CommonKeyObject[]>>;
-  HederaTokenKeyValueType: IHederaTokenServiceKeyValueType[];
-  setChosenKeys: Dispatch<SetStateAction<Set<IHederaTokenServiceKeyType>>>;
-  setKeyTypesToShow: Dispatch<SetStateAction<Set<IHederaTokenServiceKeyType>>>;
-}
-
-export const SharedSigningKeysComponent = ({
-  keys,
-  setKeys,
-  chosenKeys,
-  setChosenKeys,
-  keyTypesToShow,
-  setKeyTypesToShow,
-  HederaTokenKeyTypes,
-  HederaTokenKeyValueType,
-}: SigningKeyPageProps) => {
-  return (
-    <div className="flex flex-col gap-3">
-      <div className="flex flex-col gap-6">
-        {/* Add more key */}
-        <div className="flex flex-col gap-0">
-          {keys.length < 7 && (
-            <Tooltip
-              label={keys.length !== 0 && 'Add more keys'}
-              placement="top"
-              fontWeight={'medium'}
-            >
-              <button
-                onClick={() =>
-                  handleAddingOrRemovingKeys(
-                    'ADD',
-                    chosenKeys,
-                    HederaTokenKeyTypes,
-                    setKeys,
-                    setChosenKeys,
-                    setKeyTypesToShow
-                  )
-                }
-                className="w-full rounded border border-white/30 text-center text-sm hover:border-hedera-purple hover:text-hedera-purple transition duration-300 hover:cursor-pointer "
-              >
-                {keys.length === 0 ? `Add signing keys to the token` : '+'}
-              </button>
-            </Tooltip>
-          )}
-        </div>
-
-        {/* Key wrapper */}
-        {keys.map((key) => (
-          <div key={key.keyType} className="flex gap-3 items-center">
-            {/* Key Type & Key Value Type */}
-            <div className="flex gap-3">
-              {/* Key type */}
-              <Select
-                _focus={{ borderColor: '#A98DF4' }}
-                placeholder={key.keyType}
-                className="w-[120px] hover:cursor-pointer rounded-md border-white/30"
-                onChange={(e) =>
-                  handleKeyTypeOnChange(e, key, setKeys, setChosenKeys, setKeyTypesToShow)
-                }
-              >
-                {Array.from(keyTypesToShow).map((keyType) => {
-                  return (
-                    <option key={keyType} value={keyType}>
-                      {keyType}
-                    </option>
-                  );
-                })}
-              </Select>
-
-              {/* Key Value type */}
-              <Select
-                _focus={{ borderColor: '#A98DF4' }}
-                className="w-[220px] hover:cursor-pointer rounded-md border-white/30"
-                onChange={(e) => handleKeyValueTypeOnChange(e, key, setKeys)}
-              >
-                {HederaTokenKeyValueType.map((keyType) => {
-                  return (
-                    <option key={keyType} value={keyType}>
-                      {convertCalmelCaseFunctionName(keyType)}
-                    </option>
-                  );
-                })}
-              </Select>
-            </div>
-
-            {/* Key value */}
-            <div className="w-full">
-              {key.keyValueType === 'inheritAccountKey' ? (
-                <Select
-                  _focus={{ borderColor: '#A98DF4' }}
-                  className="hover:cursor-pointer rounded-md border-white/30"
-                  placeholder="Select option"
-                  onChange={(e) => handleUpdateKeyValue(e, key, setKeys)}
-                >
-                  <option value={'false'}>false</option>
-                  <option value={'true'}>true</option>
-                </Select>
-              ) : (
-                <Input
-                  value={key.keyValue as string}
-                  type={'text'}
-                  onChange={(e) => handleUpdateKeyValue(e, key, setKeys)}
-                  placeholder={
-                    key.keyValueType === 'contractId' ||
-                    key.keyValueType === 'delegatableContractId'
-                      ? 'ID of a smart contract instance...'
-                      : `${key.keyValueType.split('_')[0].toUpperCase()} compressed public key...`
-                  }
-                  size={'md'}
-                  focusBorderColor={'#A98DF4'}
-                  className={'w-full border-white/30'}
-                />
-              )}
-            </div>
-
-            {/* delete key button */}
-            <Tooltip label="delete this record" placement="top">
-              <button
-                onClick={() =>
-                  handleAddingOrRemovingKeys(
-                    'REMOVE',
-                    chosenKeys,
-                    HederaTokenKeyTypes,
-                    setKeys,
-                    setChosenKeys,
-                    setKeyTypesToShow,
-                    key.keyType
-                  )
-                }
-                className={`border h-fit border-white/30 text-base px-1 py-1 rounded-lg flex items-center justify-center cursor-pointer hover:bg-red-400 transition duration-300`}
-              >
-                <AiOutlineMinus />
-              </button>
-            </Tooltip>
-          </div>
-        ))}
-      </div>
-
-      {/* tip link */}
-      {keys.length > 0 && (
-        <small className="text-sm">
-          For comprehensive details about keys, please visit{' '}
-          <Link
-            href={
-              'https://github.com/hashgraph/hedera-smart-contracts/blob/main/contracts/hts-precompile/IHederaTokenService.sol#L116'
-            }
-            target="_blank"
-            className="text-hedera-purple underline font-medium"
-          >
-            this resource.
-          </Link>
-        </small>
-      )}
-    </div>
-  );
-};
 
 /** @dev shared component representing the list of transactions */
 interface TransactionResultTablePageProps {
